+++ changelog.md
@@ -2,6 +2,7 @@
 - [Enhancement] Add default items for the first seat into a lesser slots.
 - [Fix #178] Full mass used for empty parts for purpose of removing from containers.
 - [Fix #179] Put all KIS items into a real KSP categories.
+- [Fix #180] Persist equipped state in EVA inventories.
 
 ### 1.3.1 (November 23rd, 2016)
 - [Change] Stop using KSPDev obsolete logging methods.